+++ app/server/services/Activity/ActivityService.js
@@ -109,7 +109,7 @@
     userService.getSessionUser(token)
       .then(function (sessionUser) {
         //find all activities from user and populate participants
-        return Activity.findAllActivities({
+        return Activity.queryInActivities({
           select: '_id name story created participants status',
           where: ['_creator'],
           conditions: [sessionUser._id],
@@ -153,71 +153,11 @@
     return deferred.promise;
   };
 
-  /**
-    Insert a (new)User on Activity
-  **/
-  var insertNewUser = function (activityId, userEmail) {
-    var deferred = Q.defer();
-
-    var activitySizeBeforeUpdate = GLOBAL.CONST_EMPTY_NUMBER;
-    var queryActivityParticipants = {
-          select: '-_id participants',
-          where: ['_id'],
-          conditions: [activityId]
-    }
-
-    Activity.findAllActivities(queryActivityParticipants)
-      .then(function (activities) {
-        activitySizeBeforeUpdate = activities[0].participants.length;
-        return User.getUserByEmail(userEmail);
-      })  
-      .then(function (user) {
-        if (user !== null) {
-          var updateQuery = {
-            $addToSet: { participants: user._id }
-          };
-          return Activity.updateActivity(activityId, updateQuery);
-        } else {
-          return userService.inviteUserToApplication(userEmail, activityId);
-        }
-      })
-      .then(function (activity) {
-        if (activity.participants.length > activitySizeBeforeUpdate) {
-          deferred.resolve({ message: Message.SUCCESS_INSERTING_USER });
-        } else {
-          deferred.reject(Exception.ERROR_ACTIVITY_USER_INSERT);
-        }
-      })
-      .catch(function () {
-        deferred.reject(Exception.USER_INSERTING_ERROR);
-      });
-
-    return deferred.promise;
-  };
-
-  var removeUser = function (activityId, userId) {
-    var deferred = Q.defer();
-
-    var updateQuery = {
-      $pull: { participants: userId }
-    };
-
-    Activity.updateActivity(activityId, updateQuery)
-      .then(function () {
-        deferred.resolve({ message: Message.SUCCESS_REMOVING_USER });
-      })
-      .catch(function (error) {
-        deferred.reject(Exception.USER_DELETING_ERROR);
-      });
-    return deferred.promise;
-  };
 
   // export the class
   module.exports = {
     createNewActivity: createNewActivity,
     getTeacherActivities : getTeacherActivities,
     deleteActivity: deleteActivity,
-    insertNewUser: insertNewUser,
-    removeUser: removeUser
   };
 }());
